Turn on heap profiling for memory.desktop benchmark
The memory metric related to malloc used by telemetry is noisy
because it is based on the HeapWalk implementation and affected
by the tracing infra-structure.

A shim_allocated_size / shim_allocated_objects metric is added
to provide a more precise metric based on objects allocated
through the shim. This metric is much more precise but require
heap-profiling to be turn on.

Related to:
  https://codereview.chromium.org/2865333002
  https://codereview.chromium.org/2801153004

Note to Perf sheriffs, this CL may:
  * affect memory consumption for the 'memory.desktop' benchmarks
  * slowdown 'memory.desktop'benchmarks by adding an extra cost for tracking allocations.

R=erikchen
BUG=713191

Review-Url: https://codereview.chromium.org/2900223002
Cr-Original-Commit-Position: refs/heads/master@{#474423}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9af9d8069f1a0e3862c5c8526e20ef529f402134