Migrate GPU histograms from Google-internal repo to Chrome repo.
This should have been done years ago in order to simplify these
histograms. Doing it now as part of the cleanup of one of them.

This is a pure migration -- the histograms are unchanged except for
changing the owner.

BUG=536190

Change-Id: I02c7393e3893d22225b3e0b153a96dbeedebaf29
Reviewed-on: https://chromium-review.googlesource.com/515702
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#474807}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3a5be87dc36a4267a126c809880ed2621568cb98