Use NetworkTrafficAnnotation in URLRequestContext UMA logging
This CL makes use of NetworkTrafficAnnotation that is associated with
each URLRequest to log UMA histograms.

When 'Net.URLRequestContext.OutstandingRequests' is logged, we
additionally log what the URLRequest is created for. This hopefully
can tell us why we have a large number of OutstandingRequests in the
extreme scenarios.

BUG=711721

Review-Url: https://codereview.chromium.org/2897143003
Cr-Original-Commit-Position: refs/heads/master@{#474706}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 669c310279272d3202f2784c4323b392a52498d3