Stability instrumentation Crashpad integration
The stability instrumentation is about Chrome recording internal
state to a memory mapped file as it is running. It is currently used
experimentally on Windows only.

This CL makes use of Crashpad's extensibility mechanism to collect
information from the stability file on crash, and include it within
the crash report as an additional minidump user stream.

DETAILS:
- The browser process now passes the user data dir to the crash
handler's command line.
- The handler process creates the Crashpad extension that will search
for a stability file on any crash. If the file is found (currently only
possible for instrumented browser processes), it is collected and the
information is added to the crash report. On successful collection the
stability file is deleted.

TESTING
- launch chrome with the stability instrumentation:
  chrome.exe --user-data-dir=<data-dir> --enable-features=StabilityDebugging
- validate the presence of the stability file in <data-dir>/Stability
- trigger a crash (visit chrome://inducebrowsercrashforrealz)
- validate the deletion of the the stability file in <data-dir>/Stability
- validate the presence of a crash report in
<data-dir>/Crashpad/Reports and ensure it contains the additional
stream, either with dumpchk.exe or the dump_stability build target.

BUG=718437
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.win:win10_chromium_x64_rel_ng

Review-Url: https://codereview.chromium.org/2867063002
Cr-Original-Commit-Position: refs/heads/master@{#474759}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fecdf607e706474b65a4d26e554f0504aead3f3e