[MD Bookmarks] Fix singleton tab behavior.
This CL fixes an issue where MD Bookmarks would not focus an existing
tab when the bookmark manager was launched a second time. This was
happening because the query of the URL was being used in the deduping
process, so calling a function which sets IGNORE_AND_NAVIGATE fixes this.

It also fixes an issue where repeatedly pressing the bookmark manager
shortcut would repeatedly refresh the page because the specified URL
is chrome://bookmarks but the redirected URL is chrome://bookmarks/?id=1.

BUG=725761

Review-Url: https://codereview.chromium.org/2899223002
Cr-Original-Commit-Position: refs/heads/master@{#475290}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 870ec553a786c0f19d6785a4333fa7a6dc2885c0