Add UMA metrics for the linux distro.
At this point, only included the distros that are currently supported
both by us and the distro maker. For example, Ubuntu 15.10 is not
included as it is no longer supported by Ubuntu.

If there is a large UNKNOWN usage, then we can try to add other popular
distros, such as Mint.

BUG=724244
R=rkaplow@chromium.org,thomasanderson@chromium.org

Review-Url: https://codereview.chromium.org/2902853002
Cr-Original-Commit-Position: refs/heads/master@{#475653}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fcd78795faa1db43a347464fa235ead2a186636b