Enable the CPU time metric for system_health.common
We always intended for this metric to be enabled on system health, but
were unable to do so because enabling 'toplevel' for these stories would
blow up the trace size to the point where Telemetry or trace viewer
could no longer handle it. Now that we've fixed the problems with
Telemetry and trace viewer, it should be safe to enable this.

TBR=nednguyen@google.com

Bug: 704197
Change-Id: If5771f4eaeedd4dc34242530868a9acda059d01a
Reviewed-on: https://chromium-review.googlesource.com/520145
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#476097}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ea35356f880e66988f7f96ebf8a927677efbe58