Force off clang for GPU FYI release bots.
The GPU FYI release bots went red when we switched on clang by default.
It doesn't sound like that'll be fixed soon, so let's explicitly disable
clang on these FYI bots so that they're not affected the next time we
turn it on by default.

BUG=727437

Review-Url: https://codereview.chromium.org/2920573002
Cr-Original-Commit-Position: refs/heads/master@{#476048}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f307f2f73c304ee22ac9fb43d3ce4fbf493ae4a