Track how often we would attempt to resolve non-'compliant' domain names.
We have 2 hostname goodness predicates: checking that
`net::IsValidHostLabelCharacter` returns true for all characters in all labels,
and `net::IsCanonicalizedHostCompliant`. The latter is more strict than the
former, in that it imposes an additional constraint on the final label: it can
only begin with an alphanumeric character.

Add a counter to see how often hostnames would pass the former check but fail
the latter, stricter check. This will help us decide which predicate to finalize
on.

BUG=695474

Review-Url: https://codereview.chromium.org/2906723002
Cr-Original-Commit-Position: refs/heads/master@{#476122}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64af80e692404a380476b4b7edb27392eae75014