[Perf] Filter memory metrics in media benchmarks
Recently, memory metrics were turned on for media benchmarks, but the filter
used to reduce the volume of memory data sent to the perf dashboard was not
used. This CL adds the filter to these benchmarks to reduce the chance of the
perf dashboard being overloaded.

BUG=728203
NOTRY=true

Review-Url: https://codereview.chromium.org/2917423003
Cr-Original-Commit-Position: refs/heads/master@{#477069}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7c4b7842b0e593a4b165438a4d1281a4c006c4b0