Remove ProfileStatistics calculation on Startup and Shutdown
In order to reduce the latency on shutdown and possibly also on
startup, ProfileStatistics will not be precalculated and stored in
ProfileAttributesStorage anymore but calculated when the removal of a
profile is requested.
This requires the profile to be loaded if it isn't already loaded.

BUG=716267
TBR=tommycli@chromium.org

Review-Url: https://codereview.chromium.org/2909053002
Cr-Original-Commit-Position: refs/heads/master@{#477257}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 499d5e9e59c379dcfc4e8d782817246a43ce4130