Record resource scheduler UMA
Record UMA on how many requests are in-flight. UMA is recorded only when
the set of in-flight requests changes.

Also, record UMA on highest number of delayable requests in-flight when
a delayable or layout blocking request is in flight. This UMA is
recorded per-request (on select requests).

BUG=720783

Review-Url: https://codereview.chromium.org/2873223002
Cr-Original-Commit-Position: refs/heads/master@{#477447}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29fc63386af102f88a5373b8b83ee94ac6b4ec45