Don't initialize the BattOr before retrieving the git hash
I think this is safe, but it's not completely clear to me. We only
recently started initializing the BattOr before retrieving the git hash
(in https://crrev.com/2622003005). However, that change was attached
to a bug that had to do with retrying while retrieving the git hash,
which we now do. I can't remember if there was real rationale behind
why we made the change or if we were just trying to see if that'd fix
a problem.

All in all, neither I nor Aaron can see any theoretical reason why we
shouldn't be able to retrieve the git hash before an init, so I think
it's worth trying again now that we have retry logic in place. This
should cut the time to start tracing by 4 seconds, which is significant.

TBR=aschulman@chromium.org

Bug: 711277
Change-Id: Iaaf2a695551f4abbf1539ef047ebd48c224683c2
Reviewed-on: https://chromium-review.googlesource.com/527327
Commit-Queue: Charlie Andrews <charliea@chromium.org>
Commit-Queue: Aaron Schulman <aschulman@chromium.org>
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#477984}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1689c7e8fdfed4dc543d02e7743a3733be6b19e5