[Findit] Separate parsing OWNERS file tags and component mapping generation.
This is the first patch to implement the generated file detailed in
https://docs.google.com/document/d/1TGmugh1RGsQpHwkWA6Yo-J0_F_0vxgMjk06uDSZ5u0s/edit#heading=h.s59ky3bav81t

It separates the logic that traverses the tree and parses OWNERS files,
from the part that aggregates and validates such data.

It also adds support for the `# OS:` tag.

R=sergiyb@chromium.org,katesonia@chromium.org,stgao@chromium.org
BUG=728002

Change-Id: I930f3c32d0185165c3d6b1908b2997eed09f4e56
Reviewed-on: https://chromium-review.googlesource.com/518449
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#478424}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ea8c7e4c2540865661f8420966dc5b7ea4d34cf