Remove Page35 "crowd720_vp9.webm&seek" since not needed.
Also, removing this will help with crbug/720514 to show that the test
case itself is not the root cause of the bug. (Or if it is the root
cause, then this will fix that bug.)

crowd720_vp9.webm&seek is not needed since we have tulip2.vp9.webm&seek
and crowd1080_vp9.webm&seek. Those two page provide enough vp9 coverage
to make up for this removal.

BUG=720514,710253

Review-Url: https://codereview.chromium.org/2929003003
Cr-Original-Commit-Position: refs/heads/master@{#478341}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f243d9ff32782239f6d6c49c9c9fab2ca5021d2