Add metrics for new CBD dialog
Add metrics for the new Clear Browsing Data dialog on Android.
A histogram and actions are used to record which tab was used for a
deletion.
Time period changes are recorded as actions.
A bug is fixed where "ClearBrowsingData_DialogCreated" was recorded
twice because the new dialog creates two instances of
ClearBrowsingDataPreferences.

BUG=731120

Review-Url: https://codereview.chromium.org/2927063002
Cr-Original-Commit-Position: refs/heads/master@{#478602}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 412ef84048d767626e0d95ad2c14fd3f43398eff