Move ranker_model_loader to a new component.
This change should not incur any behavior change. It moves the RankerModelLoader code from components/translate to a new component. The ModelLoader (and other parts of Ranker) will be re-used by contextual_search (and probably other components eventually).

The plan for the machine_intelligence component is that it will hold the code for machine intelligence features in Chrome. In the short term, we plan to put the Ranker code here (go/chromeranker), including the ModelLoader, example logging and model inference.

The ModelLoader is the first part to be moved out of Translate, since it is already ready for re-use by other components (e.g. Contextual Search). Other parts will need to be refactored first before we move them out of the Translate component.

BUG=731845

Review-Url: https://codereview.chromium.org/2925733002
Cr-Original-Commit-Position: refs/heads/master@{#482754}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97b0ab02964dac76c919c24cac8b50ad1141cd24