cros: Enable Ext4 dircrypto migration for ARC kiosk mode.
When old encryption is detected in ARC kiosk session, we should migrate
the encryption. As it is for kiosk situation, we skip all checks
(user confirmation, battery, and space) and start migration immediately.
This CL also adds UMA stats for the kiosk migration.

BUG=730966
TEST=manually tested by following https://docs.google.com/document/d/1JS6wtZ6VxKDN0SvphuvvydixPkiTwlxzJ17DZUsMEzM/edit#

Review-Url: https://codereview.chromium.org/2944553003
Cr-Original-Commit-Position: refs/heads/master@{#482856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a07fb43185f855ed04446008efebf80564ef3c9