[Tab Metrics] Added TabManager.SessionRestore.SwitchToTab UMA.
This is the first of two metrics recorded in the TabManager when switching to a
different tab. If there is currently a session restore in progress and the user
switches tabs, we emit an UMA histogram recording the tab loading state. As a
follow-up, we'll add a separate UMA if the TabManager is in a background tab
loading state, once the tab manager has that functionality.

BUG=732837

Review-Url: https://codereview.chromium.org/2962773002
Cr-Original-Commit-Position: refs/heads/master@{#483525}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97e4ec20d584f59fe5a701cde8b29257b0a0b31d