[TTS] Add an Entity Heurisic for ML signaling.
Adds a new heuristic for taps that look like they are
on patterns that we guess are probably entities.  We have a
very simple recognition algorithm in this first cut that is
designed to be simple to implement and test.  It should have
high precision but low recall, and it cuts many corners.
Basically it just looks for two adjacent camel-case words for
users in English speaking countries.

Adds junit tests for ContextualSearchContext.java.

BUG=723194

Review-Url: https://codereview.chromium.org/2930533002
Cr-Original-Commit-Position: refs/heads/master@{#483598}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43a9c19cc867657522240eb117eb09edef4a95b6