[NTP::Metrics] Do not record not visible sections.
This CL deprecates CountOnNtpOpened and replaces it with
CountOnNtpOpenedIfVisible. The latter ignores not visible categories.
Previously, they were recorded as "showing 0 suggestions", which was not
useful for us.

A section is considered visible if it is present in
SectionList.mSections after a |resetSections| call.

Bug: 737602
Change-Id: I04062ae5fa672faf08742ce83de3886ce071ff99
Reviewed-on: https://chromium-review.googlesource.com/558906
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: vitaliii <vitaliii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484852}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2ca50c8131f7826a9d5c06730dcb39bdb189c346