Fix Service Worker UMAs of NEW_PROCESS
In the current implementation, when SiteInstance was used to get the
process, the Service Worker UMAs are recorded as NEW_PROCESS. But This
is not accurate. SiteInstance can reuse the existing process.

This CL also splits the EXISTING_PROCESS UMAs to EXISTING_READY_PROCESS
and EXISTING_UNREADY_PROCESS. When PlzNavigate is enabled, the service
worker is likely to be started in an existing unready process which was
newly created for the navigation. We should record this case apart from
starting in an existing ready process.

Bug: 740102
Change-Id: I523cea926fb78eb25e2669cfa0a3d56ca003e08c
Reviewed-on: https://chromium-review.googlesource.com/564912
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#485839}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 672f26e7d115ad9ef24b83f2bf9c4fbe986b0aa1