[TTS] Add onTouchDown to GestureStateListener.
Adds messaging from ContentViewCore to GestureStateListeners when any
TouchDown occurs to allow timing of Tap events for Contextual Search.

Adds a heurisitc to Contextual Search to categorize a Tap by duration.
CS will use the Tap duration as a signal for machine learning
with the hope that accidental taps are short, and deliberate taps tend
to be longer in duration.  Log histograms for short and long duration
tap CTR, and histograms of durations for seen and unseen to UMA.

BUG=723194,444114

Review-Url: https://codereview.chromium.org/2932623002
Cr-Original-Commit-Position: refs/heads/master@{#486900}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d1e08b84b36729cf3d0fee9ab77028e9af0a8d8