Adding previews information to PLM UKM
This adds four metrics to the PLM UKM for previews: "opt_out", set to true when the user hit "Show Original" on a preview;
"server_lofi", "client_lofi", and "lite_page" set to true when a preview was shown. When no preview is
shown, no metrics are added to the UKM.

Similarly, when the user does not opt out, nothing is added for opt_out.

This does not track offline previews yet, but may in the future.

The infobar posts an event to page_load_metrics for the web contents when the user clicks "show original" on the page.

BUG=701514,723711,728707

Review-Url: https://codereview.chromium.org/2952343004
Cr-Original-Commit-Position: refs/heads/master@{#487238}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1806f8e181472bb2fb3d32ce575c4b2ec8a36bf2