Add Event-prefix to histogram enum for In-Product Help
The Feature Engagement Tracker histograms were not updated at
the same time as the rest of the code was updated to use an
Event-prefix for all class names and members in:
https://chromium-review.googlesource.com/580707

This adds the same convention to the enum values used for UMA
histograms, and updates both the C++ enum, and the UMA label.

Since this is only a descriptive label, and the integer value does
not change, it is safe to edit the current value instead of adding
a new one.

BUG=744953

Change-Id: Ib7f547406041c2fe8ebdaf945de443d8b5a0138c
Reviewed-on: https://chromium-review.googlesource.com/585233
Reviewed-by: Xing Liu <xingliu@chromium.org>
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489544}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a2282243c3ad2ac00246dcbbab8ce63469daffe