[Telemetry] Make clock_domain_test run by itself.
By adding the @decorators.Isolated decorator we can ensure that
this test runs by itself. I believe the issue with this test is that
it is seeing multiple clock sync markers in the trace, which is causing
it to fail an assert. By isolating it, it should make it so that this is
no longer an issue.

Bug: 748655
Change-Id: I01bf75a31806d0ac53f976154fdbd014269e7a4c
Reviewed-on: https://chromium-review.googlesource.com/590366
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: rnephew <rnephew@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 910f8a81e3f89e13c5bf4850e1cb6bf9a3cdcebf