Enable SafeBrowsing for WebSocket handshake in Chrome
Enable the in-renderer-process implementation of SafeBrowsing when the embedder
is Chrome. It remains disabled for android_webview. When the browser-side
SafeBrowsing mojo service is not running it will treat all handshakes as safe.

Also enable the browser-process SafeBrowsing mojo service whenever SafeBrowsing
is enabled. Previously it was only enabled when the NetworkService feature was.

Bug: 644744
Change-Id: Idd4a915541a1f7e6696af8508f9add911636f70b
Reviewed-on: https://chromium-review.googlesource.com/570141
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Yuzhu Shen <yzshen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490352}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 41b895907e65a2131803d93c690f8f43b67d7d7d