_WaitForSpeech should not wait for speech output that happened in the past.
Currently, nvda_chrome_tests.py _WaitForSpeech will block for a period of
time until NVDA speaks some given text. The implementation reads output from
a log file that NVDA writes to.  This read just starts at the start of the
file and builds a list of what was said. This means that each call to
_WaitForSpeech() must contain the prior expected text.

This CL simply remembers the last offsite and when reading the file again
we ignore old stuff.  I don't think that re-reading the file is a big deal
since we're barely testing with this approach.  At some point, if it becomes
a problem, we might think of something more clever.

Bug: 742592
Change-Id: I47b417f7aa7b7cf8f6f6361d4581ab0ea562a042
Reviewed-on: https://chromium-review.googlesource.com/599373
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Commit-Queue: Doug Turner <dougt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491869}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d07794506ed64430f85c18824253897740e86e63