NVDA test method _WaitForSpeech should ignore unrelated speech.
Current, we have a method _WaitForSpeech which blocks on NVDA to speak
exactly the given set of text.

What this CL does is modifies _WaitForSpeech such that it only tests
for the ordered set of given text and ignores text that may have
happened prior to the text we're looking for.

Because of this change, I also renamed the method to _TestForSpeech


Bug: 742592
Change-Id: I37f2ff890a2f8860afac17b0011300d8fd34dd3f
Reviewed-on: https://chromium-review.googlesource.com/599607
Commit-Queue: Doug Turner <dougt@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491881}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dad723268c7ffb48d897d4d30238110b0c8c6ccb