Implement pull-to-refresh for Aura
Pull-to-refresh effect is implemented for Aura platforms. This is behind
the --pull-to-refresh flag. The overscroll controller is changed to make
vertical overscroll behavior similar to horizontal overscroll; i.e., for
a vertical overscroll to start, the page should be scrolled to the top
before gesture sequence is started.

BUG=702404
TEST=manual with --pull-to-refresh flag

Change-Id: I74e9b08182683586a66a931615ab72b525de3af4
Reviewed-on: https://chromium-review.googlesource.com/586952
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492119}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd0d2d32e5ffe9102ae14b89fcdccd21a61eab15