[tools/perf] Reenable service_worker.service_worker_micro_benchmark
Renable the benchmark on webview to see if individual pages can be
disabled instead.

This is part of the benchmark reenabling fix-it, where we're aiming to
bring the stories that are disabled in line with the stories that are
actually still failing. We're also migrating from wholesale benchmark
disabling to per-story disabling when a story fails.

NOTRY=true

Bug: 653924,713222
Change-Id: Ia1c2f0dbd9b6346bc9ea30680c20eb7d43638768
Reviewed-on: https://chromium-review.googlesource.com/606248
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#492662}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb5d43903a7bb942cefc0122d8a49689a6018448