Add logic in ssl_error_handler to render a MITM software interstitial.
The MITM software interstitial is an interstitial that will be
displayed to users struggling with SSL errors because of an antivirus,
firewall, proxy, or other man-in-the-middle software. You can read
about this interstitial in depth at go/sperigo-design-doc.

This is the first of two CLs I intend to use to check in for the bulk
of the MITM software interstitial code. This CL includes code in
ssl_error_handler.* with the logic used to scan a certificate and
determine if it was issued by MITM software as well as unit tests.
The next CL will include an extension of the SSLBlockingPage class
for the MITM software interstitial, an extension of the SSLContentUI,
and browser tests. At a high level, this is the "back end" CL, and the
UI CL is coming next.

Bug: crbug.com/751902
Change-Id: I80d1493b0ece5b8326987f87fa301dec41a3df87
Reviewed-on: https://chromium-review.googlesource.com/599012
Commit-Queue: Sasha Perigo <sperigo@chromium.org>
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493546}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2bfbbe90fe0c3f8590080dde5021260588b4e354