Kill a renderer if it creates a GuestView of a nonsense type.
A renderer sends a GuestViewHostMsg_ViewCreated message to the browser
upon creation of a GuestView and specifies the GuestView type as a
string. A misbehaving renderer can send a nonsense GuestView type
causing the browser to crash when it CHECKs that the type is valid.

We now kill the renderer for sending the bad message instead of
crashing the browser.

Bug: 751376
Change-Id: I7527bb4f47119e689a10f8ad40aa90f2d0778575
Reviewed-on: https://chromium-review.googlesource.com/612159
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494068}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36cc7b8f04a674c4ecfaa587477d73bd78531abb