Log the avatar load status for decorating the taskbar
This is a follow-up CL to https://chromium-review.googlesource.com/c/615217
which is supposed to fix the taskbar mis-badge bug. It should be worthwhile
to log the avatar load status for a short period to see if we missed anything.

Bug: 536553
Change-Id: I1636a7eb4cc441d245f94c546ee43885c8745a60
Reviewed-on: https://chromium-review.googlesource.com/618306
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Anthony Vallee-Dubois <anthonyvd@chromium.org>
Commit-Queue: Xi Cheng <chengx@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496018}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5176671e56a9f4b335d57fc6e6b6ae33dd49a9b