Update Wikipedia story in key_mobile_sites
The recording of this was old and causing the site to "misload" when the
inert-visual-viewport feature was turned on. The recording has been
uploaded to cloud storage. This patch simply updates the hash to match
the uploaded file.

Bug: 749357
Change-Id: I2a877d5f45eab5e7597f3389dda833dc27c2b72d
Reviewed-on: https://chromium-review.googlesource.com/621295
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Ned Nguyen (ooo til 08/30) <nednguyen@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#496671}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2331c0f6e3a7a6df0883486ac70a04f2bb539e8c