[Telemetry] Use SUPPORTED_PLATFORMS to determine if a benchmark is supported on a given platform.
This is not to be confused with StoryExpectations, which is how we disable at
the story level. This is used to disable the entire bencmark when that benchmark
was never intended to run on the given platform (mobile benchmarks not running
on desktop for example).

This requires https://codereview.chromium.org/3001873002/ to roll first.

Bug: 713222
Change-Id: I37a5d57801e3dab452ef90cbb811097a51627d71
Reviewed-on: https://chromium-review.googlesource.com/627086
Commit-Queue: rnephew <rnephew@chromium.org>
Reviewed-by: Ned Nguyen (ooo til 08/30) <nednguyen@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496791}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d0436aa8cf768320fdb196053ed65f6152f6094