[CrOS Tether] If Bluetooth is not present, disable Tether entirely.
Previously, we assumed that if Bluetooth was not fully enabled, then
it was simply turned off by the user. In this case, we would show some
"enable Bluetooth" UI surfaces.

However, there are actually cases where Bluetooth is not present on the
system at all, so we should not display this UI. Additionally, there
are some bugs which cause Bluetooth to be hidden (crbug.com/752664).

This CL changes our logic so that Tether-related UI is never shown at
all.

Bug: 758376, 672263
Change-Id: Id923e056524ab000e8e9901916e14b77d775c264
Reviewed-on: https://chromium-review.googlesource.com/630360
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496866}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0de4f5aed65ec8f985d1d18a793035ead8a92a4e