[WebApps] Exposes Share and Open in Chrome through Android Notification.
With this patch standalone web apps on Android will by accompanied by an
ongoing notification which exposes the website URL, discloses that the
app runs in Chrome and gives access to 'Share' and 'Open in Chrome'
actions.

Note that this feature is in its early days and is behind a flag. The
plan is to iterate on a number of details while it's available on Canary.

Screenshot of how the notification looks like as of this patch:
Android N: https://goo.gl/photos/V6CnxEWZbSXyMiUF8
Android M: https://goo.gl/photos/bxRKtzGzU5PFzTua8

Bug: 471623
Change-Id: I84860ffeab1423701abacaa13037a12836bbe3a5
Reviewed-on: https://chromium-review.googlesource.com/627522
Commit-Queue: Piotr Swigon <piotrs@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497285}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48a541dc5c667b718899a0ce8c8a57152d9156be