Fix NewTabPage.LogoShown and .LogoShownTime recording
Before this CL, .LogoShownTime was only recorded for cached logos, i.e.
it missed the case where the logo needed to be fetched from the network.
After this CL, it records the first non-null logo (if any), which was
the original intention.

This also adds a note to the description for .LogoShown explaining that
it can be recorded twice per NTP (cached + fresh), and it adds a
.FromCache vs .Fresh split.

Bug: 760496
Change-Id: I689ce352f2f6af27d833c3de7da03169a41e2d94
Reviewed-on: https://chromium-review.googlesource.com/643298
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Chris Pickel <sfiera@chromium.org>
Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498792}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d8f396c560f794c856c19b2e90c4644d3670623