Move the connection timeout from the BattOrAgent to the BattOrConnection
With the introduction of the slow flush into opening the connection,
it's possible for opening the connection to take a very long time from
the BattOrAgent's point of view. For example, if the user records a
180s trace, that trace may take a minute to stream back. If a corrupted
data frame occurs at the beginning of that stream, the connection will
be reopoened after a 2s delay, resulting in 53s of useful flushing
before the connection will successfully open. As long as this flushing
finds bytes, Open() shouldn't time out.

We avoid this by moving the timeout into BattOrConnection, where we
can verify that flushing is continuing to find bytes.

TBR=rnephew@chromium.org

Bug: 757508
Change-Id: I33bc2e2c145ec1f8803eda4b4c14b35d699206a9
Reviewed-on: https://chromium-review.googlesource.com/644136
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499284}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73a3e594ada44c03b4bd0e011c142f9d0b284dbb