[ModuleDatabase] Fix a racy condition in loaded module events
When a module load event is received, the in-memory PE image is parsed
to find the TimeDateStamp. It was possible for a module to unload before
parsing the now invalid PE image.

This is now done inside a Structured Exception Handler to handle the
possible EXCEPTION_ACCESS_VIOLATION, and there is a new fallback
mechanism that reads the TimeDateStamp from the file on disk if
necessary.

Bug: 751845
Change-Id: I7d0e16b8b55ed0ef046c52d632183babc4df524b
Reviewed-on: https://chromium-review.googlesource.com/602874
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499690}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 156777bedadd0cf9075c9729042be05b5cda459d