Update maps page in system health's browse page set.
In the maps page we used to wait for "loading" followed by "loaded"
events on the same button to check if the restaurants are loaded after
zooming / scrolling. This makes it brittle depending on when the
events happen. This cl changes it so that we actually wait for the
restaurant's id have been updated. This is not timing dependent and
hence more stable than earlier.

Drive by fix: Also removed an unused constant from earth's page.

Bug: chromium:763853
Change-Id: Ied029eed2869f293106cb3aaa1a39bc694d155d6
Reviewed-on: https://chromium-review.googlesource.com/663159
Commit-Queue: Mythri Alle <mythria@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502266}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 577103092661a3b6c2545e47a2681bef214fc64e