Remove automatic expansion of language list in Settings.
This is part of a large change for Language Settings.
This CL only fixes the automatic expansion of the language list, which adds the base language
whenever a locale-specific one is added.
The feature is behind a disabled flag.

Bug: 640763, 765077
Change-Id: Ie7d32c06bef0f933bf1e097a277ac063a9c2b54e
Reviewed-on: https://chromium-review.googlesource.com/670459
Commit-Queue: Claudio M <claudiomagni@chromium.org>
Reviewed-by: Dave Schuyler <dschuyler@chromium.org>
Reviewed-by: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503993}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31f0d7fc389035ee8403ecc22759715ee98042de