🏡 Display Snackbar when user triggered fetch fails.
When the user triggers a fetch through the more button, if the fetch
fails, display a Snackbar. Ideally in the future we can offer the user
information specific to how it failed and actions to remedy it.

Bug: 663376, 765603
Change-Id: I12fb27631454fb015803ae4a24a3d69c0a13357c
Reviewed-on: https://chromium-review.googlesource.com/671269
Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Commit-Queue: Peter Conn <peconn@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504427}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 75ce6b199c79a69bab775086cbabadea34a07017