ax_dump_events
Standalone executable tool that helps developers debug IAccessible2 events from any application.
This tool is an alternative to AccProbe, which is buggy, and AViewer, which is not feature rich.
The first version of ax_dump_events is barebones and simply displays events on console, but has
a lot of room for future expansion (e.g., serve info to a GUI, dump the tree structure etc.).

As an example of how this makes easier, when there is a bug with NVDA/JAWS in Chrome that doesn't 
occur with Firefox, we can check what events are emitted by Firefox for the same interactions and 
compare them with Chrome.

The ax_dump_events tool shares code with our testing infrastructure (AccessibilityEventRecorder). 

Run via:
> ax_dump_events --pid=[processid]

Bug: 768984
Change-Id: I4035dfb45ab5bb78bb4cd55f3f31c332fded5f99
Reviewed-on: https://chromium-review.googlesource.com/685694
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504667}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77c82741f2b490b01bb2dca1d1717ae9325abf17