[Home] Add promo/switch dialog for Chrome Home
This change adds a promo dialog and flag to enable it on every
restart. Choosing a setting different from the one the browser is
currently in will cause the browser's activity to restart when the
dialog is closed. The enabled state of the browser when the promo is
in use is driven off of a new shared preference
"chrome_home_user_enabled". The new preference takes priority over
about:flags.

BUG=767738

Change-Id: I4d0b5da6cf47baf338dbbc2fd8ffb85ea9140211
Reviewed-on: https://chromium-review.googlesource.com/683248
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505096}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4214d2a134c8a512e748198c8743bd54a3676d51