Support filters in file metrics provider.
Sometimes it is necessary to delay or skip the processing of certain
files (perhaps because it is still being written). Filtering gives
these options to the call site.

Moving all the "source" parameters into a single structure and passing
that allows for more fields to be added in the future with default
values so they don't have to be specified by callers that don't need
to change them.

Bug: 696721
Change-Id: I3bbfcc3a05ceba9d37e32ae86406599328554b9e
Reviewed-on: https://chromium-review.googlesource.com/678453
Commit-Queue: Brian White <bcwhite@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505182}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0cfe081d6671fbb39d1d0163138aa142e5863ac