[Home] Add personalized suggestions flag
Adds a feature that enables personalized suggestions when Chrome Home is
enabled. This value is then stored in a preference and read in
omnibox_field_trial.cc to determine whether personalized suggestions
are enabled.

Chrome Home has rather complex logic dictating whether or not it is
enabled, including an in-flight opt-in/opt-out that overwrites the base
feature value. Due to this, we cannot rely on a simple feature value in
omnibox_field_trial.

BUG=767273

Change-Id: I00e70c525244dd82e5f811ceeeab6a14c4e49994
Reviewed-on: https://chromium-review.googlesource.com/683295
Commit-Queue: Theresa <twellington@chromium.org>
Reviewed-by: Maria Khomenko <mariakhomenko@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505624}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd71c5c61558e085a2f775d85e411647bcc3dcc4