Chrome Cleaner: track the active page on reboot required
We may not be seeing as many reboots in the new flow because
there is no browser window open when the cleaner finishes. If
that's the case, we could add logic to reopen the Settings
page when a window pops in the foreground. However, I'd rather
make sure this problem exists before adding more logic to an
already complex piece of code.

Bug: 
Change-Id: I03deaadcda6e2a85001e360811c3fd53554d71a1
Reviewed-on: https://chromium-review.googlesource.com/692695
Commit-Queue: Fabio Tirelo <ftirelo@chromium.org>
Reviewed-by: Chris Sharp <csharp@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506390}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69bfccc68afeaf51c5d6128d49e61a8c2f8a292d