Add a network traffic annotation presubmit checker.
We require every time Chromium talks to the network to be explainable.
In order to do this, we annotation functions that call out to the network.
This CL adds the tooling to ensure that all of the functions are correctly
annotated and that the annotations are syntactically correct and
semantically complete.

Bug: 690323
Bug: 656607
Change-Id: I3e5a8a3d753b70768dd42ad4aa880087a089b61d
Reviewed-on: https://chromium-review.googlesource.com/592721
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 099b747b3cfc0e1788e9997784d784b3380bb045