[Background Fetch] Introduce |unique_id| to distinguish registrations.
Developers pass in an DOMString "id" when creating or getting a
Background Fetch registration, henceforth referred to as |developer_id|.

Unfortunately, these |developer_id|s can't easily be relied on as
primary keys, since it's possible for JS to abort a
BackgroundFetchRegistration (or wait for it to complete) then create
another one with the same |developer_id|, whilst holding onto a
reference to the previous BackgroundFetchRegistration (or ActiveFetches
or SettledFetches) objects.

The methods on all of these objects must continue to work, i.e. stale
data must not be deleted until the last object that refers to it is
garbage collected, and aborting an old BackgroundFetchRegistration must
not abort a newer one with the same |developer_id|, etc.

So instead, this patch generates a globally unique |unique_id| string
for every Background Fetch registration, and uses that as our primary
key throughout the Background Fetch code (from the Blink
BackgroundFetchRegistration objects which internally store their
|unique_id| to disambiguate stale vs active ones, through the persistent
storage layer, all the way to the Delegate that talks to the
DownloadService and will eventually group notifications according to
their |unique_id|).

This will ultimately make the code a lot simpler, as consistently using
these |unique_id|s also removes various race conditions between
Background Fetch registrations completing/aborting and new ones being
created with the same |developer_id|.

(And if we ever moved to global storage instead of or in addition to
per-ServiceWorkerRegistration storage in the SWDB, |unique_id|s would
have the added benefit of being globally unique, so we wouldn't need to
concatenate them with the profile ID and service_worker_registration_id
in order to use them as keys in a global map).

Bug: 757760
Change-Id: I33ddf6f42fcc04de3cfcf8e60c286b6072182274
Reviewed-on: https://chromium-review.googlesource.com/693099
Commit-Queue: John Mellor <johnme@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d5e912d98b0eaa553d4e4c3f4d02ad52c157f1c8