mac: Build lld on Mac too, and upload it to a separate zip.
We don't use lld in mac builds, so it shouldn't be in the main zip, but it's
useful for mac->win cross builds.  So upload it in a separate zip, so we can
download it when targeting win on a mac host.

Bug: 495204
Change-Id: I3733b9bea9a9326a434d5f5ff63fb0b978b950c8
Reviewed-on: https://chromium-review.googlesource.com/706813
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507408}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 251622585cfd0f0a277aa8041fa583ca50b2a854