Add parameters for limiting file age and directory size.
Allow source parameters that limit the number of files, and the total
number of bytes in those files, in a directory.  Also set the maximum
allowed age of a file that is to be processed.  These are all optional
parameters.

Bug: 760317
Change-Id: I2a1af2afb7f82cf0a72e431ba97ca57ac0393cac
Reviewed-on: https://chromium-review.googlesource.com/698665
Commit-Queue: Brian White <bcwhite@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508070}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7102c9c058a0b3b5673876447c20ab6e98a7614