[Telemetry] Remove references to ShouldDisable from benchmarks.
This is now accomplished by StoryExpectations. The few benchmarks that were
disabling based on battor detection are safe to delete because we can use
other metrics such as cpu percentage as proxies if battors are not present, so
they are still useful to run.

Bug: 713222
Change-Id: I7f793efe4294ee582586ce875d5030e33ebec2b7
Reviewed-on: https://chromium-review.googlesource.com/710215
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: rnephew <rnephew@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508215}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d009b6579f1c2dc9c784e65f027ef78031f0f07