Add support for tracking_only feature engagement configuration.
It was hard to compare experiment groups for in-product help rollouts,
since the tracking groups would never track their trigger state, since
the configuration was crafted such that the tracker would never return
true when asking to show in-product help.

This CL adds a new configuration field |tracking_only|, which when set
to true ensures that Tracker::ShouldTriggerHelpUI(...) will always
return false, but if all other conditions are met, it will still be
recorded as having been shown in the internal database and through UMA.

This is meant to be used by either local tests or for comparisons
between different experiment groups.

If one wants to later transition users with this flag set to `true` to
in fact display in-product help, they should use a different
`EventConfig::name` for the `event_trigger` configuration than the
non-tracking configuration.

BUG=761084

Change-Id: I8aefb85465eeb8ac4be24c5730c71ce2d9d2b961
Reviewed-on: https://chromium-review.googlesource.com/707636
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508379}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf591b593a17965b6fb2ea84413e909914f31675