Fix cfi-icall build by expanding the CFI blacklist
This adds a wide blacklist for failing cfi-icalls that will pass all
unit tests on Linux. This is functionally a to-do list of libraries and
routines that need to be fixed or refactored to work with cfi-icall.

Bug: 772906
Change-Id: Ie483a818f95b594d799622b52a13086686522c25
Reviewed-on: https://chromium-review.googlesource.com/719733
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: Peter Collingbourne <pcc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#509633}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3d8ae2fdc149b6bb42dfe30c1cb5b1d8dfe6a16a