Cancel the BattOr agent action timeout as soon as a message is read
Before this change, the timeout callback was only cancelled if the
message was successfuly read. This caused problems when the message
*wasn't* successfully read: the action would be scheduled to be retried
in two seconds, but before that two seconds elapsed, the timeout would
trigger and we'd say that the action failed due to a timeout, without
ever having retried it.

Cancelling the timeout unconditionally should be the right thing to do
here: after all, the timeout is intended to prevent nothing from
happening for an extended period of time and, in the case that a
bad message was read, something *did* happen, it just wasn't good.

Bug: 775520
Change-Id: I9ac7c4cfb9f12e50af07a1f8011a2495447a2f40
Reviewed-on: https://chromium-review.googlesource.com/726304
Reviewed-by: rnephew <rnephew@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#509863}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 67b7c1a18515f034ba788e055f8218c3676a8cf2