[tab_under] Simplify visibility metrics and add a new one
We used to have two visibility trackers, one for the tab in general,
and one for tabs after they did a tab-under.

This doesn't actually make a whole lot of sense, as one of the trackers
is just a subset of the other one! This patch removes the tab-under
visibility tracker in favor of just logging a TimeDelta instead.

We also add a new metrics for visibility time *before* the first
tab-under, for tabs that do tab-unders.

Bug: 661629
Change-Id: Ia54cc552e5aba131d3a8e63fad7a4f7d99836a85
Reviewed-on: https://chromium-review.googlesource.com/730857
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#510664}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 386b20f3058097507fe3600f6f0e0325848af936