clang plugin: improve handling of virtual / override fix it hints
Removing virtual is tricky because the AST does not provide the location
of the virtual keyword. Similarly, the AST also does not expose the
location of '=' in |void F() = 0| or |void ~T() = default|.

The plugin previously tries to hack around this in two different ways:
1. for virtual, it assumes the token at |method->getLocStart()| is the
   token that needs to be removed.
2. for override, it scans tokens backwards to try to find a ')' or '=',
   and guesses the appropriate insertion point.

These heuristics break down sometimes, such as when a destructor is
defaulted. This patch cleans up the token scanning logic so it works for
the aforementioned case, and uses the same logic for determining removal
of virtual as well.

This also fixes the plugin to:
1. correctly prioritize retaining the 'final' specifier over the
   'override' specifier
2. when generating the 'override' specifier fixit, also generates a
   removal fixit for 'virtual' if needed so cleanups don't reuqire two
   passes

Bug: 777575
Change-Id: If7f9df09d3475b4e54596610f2eb5fb477e5cc1a
Reviewed-on: https://chromium-review.googlesource.com/735079
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#511206}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d514493d6354b2d929e63ade437931cf82d6329