[Android WebAPK] Record metric with age of "pending update request" file
This CL adds a metric which records an approximation of how old the
"pending
update request" file is. The purpose of the metric is to assess how
often Chrome is killed by the OS during an update.

Bug=713655

Change-Id: I4f6177ec1be6ddb3e6321f9a79588cb8a36c6ec2
Reviewed-on: https://chromium-review.googlesource.com/671149
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#513230}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 818bcea97287ca28a736c3075fee3eef3a0a1576