[PageLifecycle] Add UMA to track when a stopped renderer is resumed
This CL adds UMA metrics to track the states of a renderer when it is
backgrounded.

Once it is backgrounded, it can either resume or stop. If it is stopped
it can also later resume. We are interested in the later case but for
completeness we track all 4 cases.

Bug: 779792
Change-Id: Iee397393205ae4de56a7e88a4d4737274b92d79b
Reviewed-on: https://chromium-review.googlesource.com/747863
Commit-Queue: Fadi Meawad <fmeawad@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Shubhie Panicker <panicker@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#513974}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65eacad0db1cd6b43f7a8df784477f9b180ea825