Port download resumption to BackgroundTaskScheduler
Move the DownloadResumptionScheduler to use BackgroundTaskScheduler
instead of GCM.  This maintains the same behavior as the old mechanism,
but just relies on BackgroundTaskScheduler to handle the job scheduling
and execution.  This also moves the scheduling decision from
DownloadNotificationService to DownloadResumptionScheduler (to be moved
to a final location after jming@'s refactor) so that rescheduling after
OS upgrade is still possible.

BUG=766847

Change-Id: If2b703adaa40550a780063d7fd60f6becdde53fe
Reviewed-on: https://chromium-review.googlesource.com/611413
Commit-Queue: David Trainor <dtrainor@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#514502}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88e4d5f0da7234775933a28336f403c5d7306809