Add metric to capture the time between request initiation and start
This CL adds a metric to measure the time spent from URLRequest creation
and when Start() is called on it. The purpose is to measure time spent
in various resource handlers before sending the request out to the net
stack.

Bug: 779914
Change-Id: I8e462a7d9fb541de38589aeb926f81047a735d50
Reviewed-on: https://chromium-review.googlesource.com/771549
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#517131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa65eacee6f10757a1fa35faefd890553834eafe