Add support for locking display of in-product help.
Up until now, there has been no way to ensure that no in-product help is
displayed during important user journeys. It can conceptually pop up at
any time.

This CL therefore adds a way to acquire a display lock that locks all
display of in-product help. It does not remove currently displaying
in-product help, but it does ensure nothing new is displayed.

When the protected user journey has finished, the lock must be
released to ensure that other in-product help can be displayed.

On most platforms, the lock is released simply by deleting the
DisplayLockHandle, which is given out as an std::unique_ptr. On Android,
there are two APIs, one in C++ (shared with the other platforms), and
one in Java. Since Java does not support destructors, a particular
release() method must be invoked on the DisplayLockHandle instead to
ensure that the C++ object is deleted.

These locks are tracked separately in UMA for why in-product help has
not been displayed: TriggerHelpUIResult.FAILURE_DISPLAY_LOCK.

BUG=776168

Change-Id: I388fe8d626a59c5128481d5758b3bb66821e0aa3
Reviewed-on: https://chromium-review.googlesource.com/758631
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#517591}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c05089f66f28c6e7e07dd3000395ae68b513dd18