Adding a traffic annotation presubmit checker for Windows.
To provide transparency on Chromium's requests to the network, we add
network traffic annotations to all functions that call out to the
network.
This CL adds the tooling required on Windows to ensure that all of the
functions are correctly annotated and the annotations are syntactically
correct and semantically complete.

Bug: 690323
Bug: 656607
Change-Id: I56ca754cf320db62cacf954c8e0bc64db2ce301a
Reviewed-on: https://chromium-review.googlesource.com/779139
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#518173}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b7e55d99e5d4567684e7caa83a81500938eda17e