android: Fix Startup.FirstCommitNavigationTime histogram
This CL makes following changes to the
Startup.FirstCommitNavigationTime3 histogram:
* Do not record the histogram multiple times per session
* Start recording from the application start time instead of the
  foreground time
* Record the histogram only for ChromeTabbedActivity
* Ignore foreground state, there will be separate histogram for this
  purpose
* Discard error, same document and non http(s) pages

The CL also renames the histogram to
Startup.Android.Experimental.Cold.TimeToFirstNavigationCommit

The rationale behind the recording the histogram from the application
start time instead of the foreground time:

The current metric doesn't cover the important part of the startup
latency because the ForegroundStartTime is measured after the native
library is loaded. That's a lot of important data missed. We're
exploring various ways to speed up the startup including the time
of native library initialization. We want to have a metric that will
show us these improvements.
Initially, this metric was measuring cold startups from the application
start but then it was changed. We would like to return to the initial
meaning. pasko@ described an evolution of the metric in
https://goo.gl/NyvX39

Bug: 778302
Change-Id: I01d7a6a73fbc504e31c1e877b1253fcf69a54ad8
Reviewed-on: https://chromium-review.googlesource.com/788875
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Alexandr Ilin <alexilin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#521771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8be495bb2adb20e7b74309ae213796c848d4526d