[tab_under] Add click through metrics for desktop
This CL abstracts some logic for "click position" to be shared by
both the popup tab helper, and the framebust tab helper.

Additionally, it stores a closure on the framebust tab helper to be
associated with each individual entry. This is used to collect metrics
for tab-under blocking on desktop.

Bug: 661629
Change-Id: Ibdc043c1d3068c38dad304be42b2d48133b3f21e
Reviewed-on: https://chromium-review.googlesource.com/809098
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#522171}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5179de1579a818b685d09934d32d137b6bb376fb