Reland: Media Engagement: Histograms for preloaded list.
This adds two histograms to MediaEngagementPreloadedList. The first is
used when the list is loaded and will allow us to see if the loads are
failing and why. The second is used when the list is checked and will
allow us to see if the list has not been loaded or is empty, as well
as how often the list is matching.

BUG=787464

TBR=jwd@chromium.org

Change-Id: Ibc24ffcf2c18856297be4cc5017d3b937257967c
Reviewed-on: https://chromium-review.googlesource.com/813754
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#523395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6b9d91ba3cd214737889a8423a2c616d0687ff5