[CrOS Tether] Add a timeout for sending a message over BLE.
The Bluetooth APIs accept callback and error callback parameters when
sending a message, but we've recently discovered that there are some
bugs which can prevent these callbacks from being invoked. This CL adds
an extra timeout which ensures that even if the callbacks are not
invoked, BluetoothLowEnergyWeaveClientConnection handles this case
robustly.

This is not the full fix to https://crbug.com/803677, but it is a
partial fix which can avoid some nasty side effects for Instant
Tethering.

Bug: 803677, 672263
Change-Id: I982c08dfe726a3c3e3164aac87ca8bec10b68be7
Reviewed-on: https://chromium-review.googlesource.com/877339
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#531062}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe22dec1fcd295c8681a6bf6c698c2469ceb3c90