[css-align] Add use counter for the 'legacy' alignment
The CSS WG resolved in May 2017 to replace the 'auto' keyword by
'legacy' in the CSS syntax of the justify-items property.

https://github.com/w3c/csswg-drafts/issues/1318

According to the spec, the 'auto' (now 'legacy') value applied to the
justify-items property should be resolved, during the style resolution
phase before layout, with its inherited value if this includes the
'legacy' keyword (plus 'left', 'right', or 'center').

There is no behavior change in the affected property due to this syntax
change.

Since 'auto' (now 'legacy') is the default value of the justify-items
property, this change has a relative big impact. However, we are only
interested on evaluate the use of the 'legacy' alignment, which has to
be explicitly set in the parent element.

Change-Id: Id2ae6735fb28ac44062aaec79828ce97d7edd5d6
Reviewed-on: https://chromium-review.googlesource.com/886365
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#532380}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1fe575dbeb077ce496b11c05bd92aed93289831