Allow Display UI for regional locales too.
Currently some locales cannot be selected by the user as display UI,
but this leads to a confusing experience because there are UI locales
that are very similar (i.e. 'English' cannot be selected, but 'English
(USA) can).
This change fixes that. We increase the set of locales that can be
selected and store the user selection in the Translate Prefs. Every
time we read the Pref, we convert locales to their corresponding
fallback if applicable.

Bug: 719435
Cq-Include-Trybots: master.tryserver.chromium.linux:closure_compilation
Change-Id: I1e41cb7d3a39160fbfb05571a573c1ebed2a7f91
Reviewed-on: https://chromium-review.googlesource.com/838726
Commit-Queue: Claudio M <claudiomagni@chromium.org>
Reviewed-by: Michael Martis <martis@chromium.org>
Reviewed-by: calamity <calamity@chromium.org>
Reviewed-by: Elliot Glaysher <erg@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Vitaly Buka <vitalybuka@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#533177}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0a80bb7df2aafcf92f942bc2fe9dc3879fb7e95