Add AppContainer ACEs to Application folder during installation.
This CL adds the AppContainer capability SIDs for chromeInstallFiles and
lpacChromeInstallFiles to the Application directory during installation.
The purpose of this change is to support the new AppContainer sandbox
which will require a second SID to grant access to the main executables.

Bug: 807249
Change-Id: I6a42451bee93c574fa293c9d4459f929a1d9cc1b
Reviewed-on: https://chromium-review.googlesource.com/897649
Commit-Queue: James Forshaw <forshaw@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#534500}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3246b3042dec724733164256d0218d4547ed1b9