cygprofile: Change error logging to IMMEDIATE_CRASH.
If this error is hit during static initialization --- which is
probable if the sentinel functions are out-of-order, calling
LOG(FATAL) can deadlock.  Instead we crash and hope that the system
will give some clue to the problem.

Bug: 758566
Change-Id: Ia48eaf9d636a83864073375977b41a8dd6f1dfe2
Reviewed-on: https://chromium-review.googlesource.com/904243
Commit-Queue: Matthew Cary <mattcary@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#534992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0162a9b1e7d6902134b1d4d12a8aa33722a84b8d