Disallows any Previews by default on Mac and Windows platforms
We have a new kill switch for all previews and since we never use
previews on Mac and Windows, this sets the default for previews to
be disabled on those platforms (actually, it only turns on for
android and linux). Linux is on as a developer debug convenience
and to be able to verify other feature control logic.

Bug: 803221
Change-Id: I341e8f2ffb75be274755a013e6d14f1728b863e9
Reviewed-on: https://chromium-review.googlesource.com/905522
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Doug Arnett <dougarnett@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#535194}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0319b6f12fc13f70bea41ebb4b8b0252253ae33