[Downloads location] Include file name/location editing on interstitial.
This is part of a set of changes to make it possible for users to choose
where they want to save their files during downloads on Android. This CL
updates the interstitial download location dialog to include the option
to edit the name and location of the download at the time of download.

This CL includes the following functionality:
- Option to change the name of the download file.
- Option to change the location of the download file through the
download directory preference.
- Updating the location of selected download file on the dialog.
- Checking the "Don't show again" box the first time the interstitial
is shown.
- Saving the file with the new name and location.

Mocks: go/downloads-location-mocks

Screenshot:
https://jming.users.x20web.corp.google.com/www/downloaddialog.png

Bug: 792775,810198
Change-Id: I997fb394e5cbbc2481a3809214b9d9aa5dc1c109
Reviewed-on: https://chromium-review.googlesource.com/896866
Commit-Queue: Joy Ming <jming@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#535839}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 41055bd7bddaa11af9ddfed9652e064ae5e0d311