Revert "Revert "Add Select-to-Speak to accessibility menu in status tray.""
This reverts commit f251a33e51ab0c4b757c6f1bfafb67db628854ef.

The failing tests in linux-chromeos-dbg have been fixed, re-submitting
the original change with that fix.

The fix is to clamp the opacity in AccessibilityLayer::SetOpacity to 0
if it is below 0. The bug seemed to have been floating point math error.

TBR=dmazzoni@chromium.org,tbarzic@chromium.org,jamescook@chromium.org,isherman@chromium.org,megjablon@chromium.org

Bug: 804506
Change-Id: I51a6e56a1814b054722f35672fe5309fe3609de6
Reviewed-on: https://chromium-review.googlesource.com/914088
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#536178}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c1e4da67bd310563da252e6db49792c05baf1ae