Add uma metric for bookmarks JSON file size and time to write
- Bookmarks JSON file size is instrumented when it's read (at startup).
This represents the whole populations of bookmarks users.

- Time to write is instrumented in ImportantFileWriter.
Metrics in ImportantFileWriter are reported with suffix
"BookmarkStorage" in our case. And accordingly, we can get accurate
metrics values for time to write bookmarks file.
This is instrumented at the write time. Bookmarks are written after
an update in the bookmarks and hence the metric will be biased by
bookmarks that are updated frequently. This is an intended
behavior though in order to assess the impact of increasing the size
of the JSON file.


Bug: 808439
Change-Id: I7570727931e52b9f7dcdd5a48fd812faf2c4b6ab
Reviewed-on: https://chromium-review.googlesource.com/901422
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#537634}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3df2894f790e335c4947e47f19dc1526f34b96a4