Track Site Engagement scores by SecurityLevel
Adds a new PageLoadMetricsObserver to track the Site Engagement score
for a visit to a site, separated by SecurityLevel. Adds UMA collection
for the final score and the change in the score between the page
committing and leaving the page, in a separate histogram for each
security level.

Additionally, adds UKM collection for the initial score, the final
score, the initial security level, and the final security level of the
page, collected on leaving the page.

These will be used for a study of reducing the HTTPS indicator UI in
the location bar.

Bug: 803544
Change-Id: Icf3500e9c0c6d8b546da9e4544cb2d36a1a3fd50
Reviewed-on: https://chromium-review.googlesource.com/902666
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Commit-Queue: Christopher Thompson <cthomp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#537947}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b49d5340cb51ba1cf5c2914a8df6bd6161045f98