Set automatic redirect to chrome://connection-help
Added a connection_help_tab_helper that monitors if a navigation to the
connection help page on the help center results in an interstitial, in
which case it redirects to chrome://connection-help. For now this only
happens if the BundledConnectionHelp feature is enabled.

Bug: 812036
Change-Id: Ic299b6594537c82f5f0d36842da72d4b5fc569c3
Reviewed-on: https://chromium-review.googlesource.com/917464
Commit-Queue: Carlos IL <carlosil@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#538205}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 55ef4b4d90198c9fc2e8598bdbb42a7d4b856dd5