Display an ambient badge for sites advertising a native app installation.
This CL shows ambient badges for sites offering native app banners when
the ExperimentalAppBanners feature is active. This is in order to
retain the ability to subtly advertise a native app and disincentivise
door-slam app install interstitials following the conversion of app
install banners to user-triggered modal dialogs.

With ExperimentalAppBanners active, no app banners are automatically
shown. Instead, sites that are eligible for banners will display an
ambient badge (currently a mini-infobar on Android) that triggers the
modal banner when tapped. This is a slight behaviour change for the
ambient badge (it previously triggered add to home screen directly;
now it triggers the banner). This allows users an additional
confirmation step before installing a native app, and keeps consistency
between web and native.

The site may also use the JavaScript beforeinstallprompt API to
trigger the app install banner directly. A follow-up CL will dismisss
the ambient badge when the app install banner is triggered in this way.

Search has been downranking sites showing app install interstitials, and
app install banners have been advertised as an alternative. With the
conversion of app install banners to modal, this CL retains the ability
for sites to use the same mechanism for app install banners.

This CL also renames the PwaAmbientBadge*Android classes to be called
InstallableAmbientBadge*. The Android suffix is removed as the infobar
header is included and compiled on non-Android platforms in order for
AppBannerManager to fulfil the
InstallableAmbientBadgeInfoBarDelegate::Client interface, which is
needed to avoid a duplicated WeakPtrFactory on AppBannerManagerAndroid.

BUG=782120
TBR=estade@chromium.org

Change-Id: Ic2f0244eabf4b57d5aa3613b92d815bb557b83bb
Reviewed-on: https://chromium-review.googlesource.com/923250
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#538263}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a32333e7b6248fe973ac434d918b83226acf138d